requirements-vercel.txt
flask==2.3.2
flask-cors==4.0.0
gunicorn==20.1.0
google-cloud-bigquery==3.11.4
numpy==1.26.3
pandas==2.1.4
google-api-python-client[drive]==2.97.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-auth==2.22.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dateutil==2.8.2
python-dotenv==1.0.0
